Extract test machinery to its own file and implement a few tests
