+++ site/content/src/main/exercises/stdlib/ByNameParameter.scala
@@ -10,7 +10,7 @@ object ByNameParameter extends FlatSpec with Matchers with exercise.Section {
 
   /** `() => Int` is a Function type that takes a `Unit` type. `Unit` is known as `void` to a Java programmer. The function returns an `Int`. You can place this as a method parameter so that you can you use it as a block, but still it doesn't look quite right.
     */
-  def takesUnitByNameParameter(res0: String) {
+  def takesUnitByNameParameter(res0: Either[Throwable, Int]) {
     def calc(x: () => Int): Either[Throwable, Int] = {
       try {
         Right(x()) //An explicit call the x function
@@ -28,7 +28,7 @@ object ByNameParameter extends FlatSpec with Matchers with exercise.Section {
 
   /** A by-name parameter does the same thing as the previous koan but there is no need to explicitly handle `Unit` or `()`. This is used extensively in scala to create blocks.
     */
-  def byNameByNameParameter(res0: String) {
+  def byNameParameter(res0: Either[Throwable, Int]) {
     def calc(x: => Int): Either[Throwable, Int] = {
       //x is a call by name parameter
       try {
