+++ site/content/src/test/scala/exercises/stdlib/AssertsSpec.scala
@@ -1,73 +1,59 @@
-package stdlib
+package exercises
 
-import shapeless._
-import shapeless.ops.function._
+import stdlib._
+import shapeless.HNil
 
-import org.scalacheck.{ Prop, Arbitrary }
-import org.scalacheck.Gen
-import Prop.forAll
+import org.scalatest.Spec
+import org.scalatest.prop.Checkers
 
-import org.scalatest._
-import org.scalatest.prop._
-import org.scalatest.Matchers
-import Matchers._
+class AssertsSpec extends Spec with Checkers {
+  // FIXME: get rid of this
+  import org.scalacheck.Shapeless._
 
-import org.scalacheck.Shapeless._
-
-class AssertsSpec
-    extends Spec
-    with Checkers {
-
-  class AClass extends Spec with Matchers {
-    def aMethod(res0: Int, res1: Int) = {
-      res0 shouldBe 1
-      res1 shouldBe 2
-    }
+  def `scalatest asserts` = {
+    check(Test.success(Asserts.scalaTestAsserts _, true :: HNil))
   }
 
-  def success[F, R, L <: HList](method: F, answer: L)(
-    implicit
-    A:     Arbitrary[L],
-    fntop: FnToProduct.Aux[F, L ⇒ R]
-  ): Prop = {
-    val rightGen = genRightAnswer(answer)
-    val rightProp = forAll(rightGen)({ p ⇒
-      try {
-        fntop(method)(p)
-        true
-      } catch {
-        case _: Throwable ⇒ false
-      }
-    })
+  // FIXME: depends on #259
+  // check(success(Asserts.booleansAsserts _, HNil))
 
-    val wrongGen = genWrongAnswer(answer)
-    val wrongProp = forAll(wrongGen)({ p ⇒
-      try {
-        fntop(method)(p)
-        false
-      } catch {
-        case _: Throwable ⇒ true
-      }
-    })
+  def `values asserts` = {
+    check(Test.success(Asserts.valuesAsserts _, 2 :: HNil))
+  }
+}
 
-    Prop.all(rightProp, wrongProp)
+class ByNameParameterSpec extends Spec with Checkers {
+  // FIXME: get rid of this
+  import org.scalacheck.Shapeless._
+
+  def `takes unit by name parameter` = {
+    val right: Either[Throwable, Int] = Right(29)
+
+    check(
+      Test.success(
+        ByNameParameter.takesUnitByNameParameter _,
+        right :: HNil
+      )
+    )
   }
 
-  def genRightAnswer[L <: HList](answer: L): Gen[L] = {
-    Gen.const(answer)
+  def `takes parameters by name` = {
+    val right: Either[Throwable, Int] = Right(69)
+
+    check(
+      Test.success(
+        ByNameParameter.byNameParameter _,
+        right :: HNil
+      )
+    )
   }
 
-  def genWrongAnswer[L <: HList](l: L)(
-    implicit
-    A: Arbitrary[L]
-  ): Gen[L] = {
-    A.arbitrary.suchThat(_ != l)
-  }
-
-  val theInstance = new AClass()
-
-  def `we can generate right and wrong props` = {
-    val aProp = success(theInstance.aMethod _, 1 :: 2 :: HNil)
-    check(aProp)
+  def `apply parameters` = {
+    check(
+      Test.success(
+        ByNameParameter.withApplyByNameParameter _,
+        "retzelpay" :: HNil
+      )
+    )
   }
 }
